🚚 move: confirm type dialog 분리 및 파일위치 변경
